requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
blinker==1.9.0
browserforge==1.2.3
cachetools==5.5.2
camoufox==0.4.11
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
Cython==3.0.12
Flask==3.1.0
fonttools==4.57.0
frozenlist==1.5.0
geoip2==5.0.1
google-auth==2.39.0
google-genai==1.10.0
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hcaptcha-challenger==0.15.3
hpack==4.1.0
httpcore==1.0.8
httpx==0.28.1
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
language-tags==1.2.0
loguru==0.7.3
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
maxminddb==2.6.3
mdurl==0.1.2
msgpack==1.1.0
multidict==6.2.0
numpy==2.2.4
opencv-python==4.11.0.86
orjson==3.10.16
packaging==24.2
patchright==1.51.3
pillow==11.2.1
platformdirs==4.3.7
playwright==1.51.0
priority==2.0.0
propcache==0.3.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pyee==12.1.1
Pygments==2.19.1
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
Quart==0.20.0
requests==2.32.3
rich==14.0.0
rsa==4.9.1
screeninfo==0.8.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.3.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.19.0
